requirements.txt
alembic==1.0.10
appdirs==1.4.3
attrs==19.1.0
black==19.3b0
Click==7.0
Flask==1.0.3
flask-blueprint==1.2.7
Flask-Migrate==2.5.2
Flask-SQLAlchemy==2.4.0
itsdangerous==1.1.0
Jinja2==2.10.1
Mako==1.0.10
MarkupSafe==1.1.1
psycopg2==2.8.2
python-dateutil==2.8.0
python-dotenv==0.10.2
python-editor==1.0.4
six==1.12.0
SQLAlchemy==1.3.3
toml==0.10.0
Werkzeug==0.15.4
